pyproject.toml
[tool.poetry]
name = "ticktick-mcp"
version = "0.1.0"
description = "TickTick MCP Server - Model Context Protocol server for TickTick integration"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "ticktick_mcp", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
mcp-python-sdk = "^1.0.0"
ticktick-py = "^1.0.0"
httpx = "^0.24.0"
pydantic = "^2.0.0"
cachetools = "^5.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
pytest-asyncio = "^0.21.0"
black = "^23.0.0"
isort = "^5.12.0"
mypy = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
ticktick-mcp = "ticktick_mcp.server:main"